take digit_sequence out from digit
