+++ crates/jpreprocess-njd/src/open_jtalk/digit_sequence/builder.rs
@@ -1,4 +1,4 @@
-use crate::{digit::symbols::Digit, NJD};
+use crate::NJD;
 
 use super::DigitSequence;
 
@@ -120,4 +120,30 @@ fn check_comma_sequence(digits: &[Digit]) -> bool {
 }
 fn check_zero_start(digits: &[Digit]) -> bool {
     matches!(digits, [Digit::Digit(0), ..])
+}
+
+#[derive(Debug)]
+pub enum Digit {
+    Digit(u8),
+    Comma,
+}
+
+impl Digit {
+    // FIXME: move this to FromStr trait
+    pub fn from_str(s: &str) -> Option<Self> {
+        match s {
+            "一" => Some(Self::Digit(1)),
+            "二" => Some(Self::Digit(2)),
+            "三" => Some(Self::Digit(3)),
+            "四" => Some(Self::Digit(4)),
+            "五" => Some(Self::Digit(5)),
+            "六" => Some(Self::Digit(6)),
+            "七" => Some(Self::Digit(7)),
+            "八" => Some(Self::Digit(8)),
+            "九" => Some(Self::Digit(9)),
+            "〇" | "０" => Some(Self::Digit(0)),
+            "，" => Some(Self::Comma),
+            _ => None,
+        }
+    }
 }
