+++ crates/jpreprocess-njd/src/open_jtalk/mod.rs
@@ -6,5 +6,6 @@
 pub mod accent_phrase;
 pub mod accent_type;
 pub mod digit;
+pub mod digit_sequence;
 pub mod pronunciation;
 pub mod unvoiced_vowel;
